requirements.txt
streamlit
scikit-learn
PyPDF2
python-docx
beautifulsoup4
requests
openai